Add guest dashboard layout
